+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support for `FilesSignature` field in packages. On mismatch it trigger
+  reinstall.
+
 ## [69.1.2] - 2021-05-28
 ### Fixed
 - Immediate reboot not being performed when combined with replan that actually